Cleanup view logic, split toolbars to separate components
