--- web_app/src/components/toolbars/ResultViewToolBar.vue
@@ -1,409 +1,0 @@
-<script setup lang="ts">
-import VButton from "primevue/button";
-import VDialog from "primevue/dialog";
-import VSidebar from "primevue/sidebar";
-import VInputText from "primevue/inputtext";
-import QuantitiesEntry from "../QuantitiesEntry.vue";
-import { useImageStateStore } from "@/stores/imageState";
-import { ImageAction, useViewStateStore, ViewStates } from "@/stores/viewState";
-import { computed, ref } from "vue";
-import { config, endpoints } from "@/config";
-import { parseClassificationsFromResponse, sendRequest } from "@/utils";
-import { type DatasetListItem } from "@/types";
-import DatasetListItemComponent from "../DatasetListItem.vue";
-import { request } from "http";
-
-// TODO: Separate to other components
-
-const imageState = useImageStateStore();
-const viewState = useViewStateStore();
-
-const quantitiesVisible = ref<boolean>(false);
-const datasetDialogVisible = ref<boolean>(false);
-const compareDialogVisible = ref<boolean>(false);
-const datasetName = ref<string>("");
-const userDatasets = ref<DatasetListItem[]>([]);
-const classifications = computed(() => imageState.objectClassifications);
-
-
-function handleReturnClick() {
-    viewState.setState(ViewStates.ImageEditPoints);
-    viewState.showBackground = true;
-    viewState.isEditingExistingResult = true;
-    imageState.clearResult();
-}
-
-
-function handleSubmitLeadersClick() {
-    submitClassificationLeaders();
-}
-
-
-function handleSubmitDatasetClick() {
-    submitDataset();
-}
-
-
-function handleCompareClick() {
-    const datasetRequestUri = config.serverUri + endpoints.getDatasets;
-    const datasetRequestPromise = sendRequest(datasetRequestUri, null, "GET");
-
-    datasetRequestPromise.then((response) => {
-        if (response.status === 200) {
-            // TODO: add thumbnails to dataset list
-            userDatasets.value = [];
-            for (const dataset of response.data) {
-                userDatasets.value.push({
-                    id: dataset.id,
-                    name: dataset.name,
-                    timestamp: Date.parse(dataset.timestamp)
-                } as DatasetListItem);
-            }
-            compareDialogVisible.value = true;
-        }
-        else {
-            console.error("Failed to retrieve datasets");
-        }
-    });
-}
-
-
-function submitClassificationLeaders() {
-    const requestUri = config.serverUri + endpoints.classifyByLeaders.replace("{image_id}", imageState.imageId.toString());
-    const requestData = JSON.stringify({
-        leaders: imageState.selectedLeaderIds
-    });
-    const requestPromise = sendRequest(requestUri, requestData, "POST");
-
-    requestPromise.then((response) => {
-        if (response.status === 200) {
-            imageState.clearResult();
-            parseClassificationsFromResponse(response.data.classifications);
-            viewState.currentAction = ImageAction.ConfirmDataset;
-        }
-        else {
-            console.error("Failed to submit dataset leaders");
-        }
-    });
-}
-
-
-function submitDataset() {
-    const classifications = imageState.objectClassifications.map((classification) => {
-        const elements = imageState.imageElements.filter((el) => el.classificationIndex === classification.index)
-        const elementIds = elements.map((el) => el.id);
-        const leaders = elements.filter((el) => el.isLeader);
-
-        if (leaders.length === 0) {
-            console.error("No leader found for classification " + classification.classificationName);
-            return null;
-        }
-        else if (leaders.length > 1) {
-            console.error("Multiple leaders found for classification " + classification.classificationName);
-            return null;
-        }
-
-        return {
-            name: classification.classificationName,
-            leader: leaders[0].id,
-            elements: elementIds
-        };
-    });
-    if (classifications.some((classification) => classification === null)) {
-        console.error("Failed to submit dataset");
-        return;
-    }
-
-    const requestUri = config.serverUri + endpoints.createDataset;
-    const requestData = JSON.stringify({
-        name: datasetName.value,
-        image_id: imageState.imageId,
-        classifications: classifications
-    });
-
-    const requestPromise = sendRequest(requestUri, requestData, "POST");
-    requestPromise.then((response) => {
-        if (response.status === 200) {
-            console.log("Dataset submitted successfully");
-            // TODO: handle response (same as registration popup)
-            imageState.reset();
-            viewState.reset();
-        }
-        else {
-            console.error("Failed to submit dataset");
-        }
-    });
-}
-
-
-function compareWithDataset(datasetId: number) {
-    const requestUri = config.serverUri + endpoints.compareToDataset
-            .replace("{result_id}", imageState.resultId.toString())
-            .replace("{dataset_id}", datasetId.toString());
-    const requestPromise = sendRequest(requestUri, null, "GET");
-
-    viewState.isWaitingForResponse = true;
-    compareDialogVisible.value = false;
-    requestPromise.then((response) => {
-        if (response.status === 200) {
-            console.log("Comparison successful");
-            imageState.clearResult();
-            parseClassificationsFromResponse(response.data.classifications);
-            datasetDialogVisible.value = false;
-        }
-        else {
-            console.error("Comparison failed");
-        }
-
-        viewState.isWaitingForResponse = false;
-    });
-}
-</script>
-
-
-<template>
-    <div class="image-view-tool-bar bar">
-        <VButton text label="Adjust" icon="pi pi-pencil" @click="handleReturnClick();" />
-        <div class="element-count">
-            <span class="element-count-value">{{ imageState.imageElements.length }}</span>
-            <span class="element-count-label">Elements</span>
-        </div>
-        <VButton v-if="viewState.currentAction === ImageAction.CreateDataset" text label="Submit leaders"
-                icon="pi pi-check" @click="handleSubmitLeadersClick" :disabled="imageState.selectedLeaderIds.length === 0" />
-        <VButton v-else-if="viewState.currentAction === ImageAction.ConfirmDataset" text label="Submit dataset"
-                icon="pi pi-check" @click="datasetDialogVisible = true" />
-        <VButton v-else text label="Details" icon="pi pi-list" @click="quantitiesVisible = true" />
-    </div>
-    <VButton v-if="viewState.currentAction === ImageAction.CompareWithDataset" label="Compare with dataset"
-            :class="(viewState.isWaitingForResponse ? 'inactive-button ' : '') + 'compare-button'" @click="handleCompareClick" />
-    <VSidebar v-model:visible="quantitiesVisible" position="bottom" style="height: auto" class="quantities" header="Counted elements">
-        <div class="quantities-label-notice">You can toggle label visibility in the settings</div>
-        <div class="quantities-header">
-            <div class="quantities-col">Count</div>
-            <div class="quantities-col">Label<span class="rename-notice">(tap to rename)</span></div>
-            <div class="quantities-col">Show boxes</div>
-        </div>
-        <div class="quantities-content">
-            <QuantitiesEntry v-for="(quantity, index) in classifications" :key="index" :index="quantity.index" />
-            <div v-if="classifications.length === 0" class="no-elements-notice">(no elements found)</div>
-        </div>
-    </VSidebar>
-    <VDialog v-model:visible="datasetDialogVisible" modal header="Submit dataset" class="dataset-dialog"
-            :dismissable-mask="true" :draggable="false">
-        <p><span>{{ imageState.selectedLeaderIds.length }}</span> categories selected</p>
-        <label for="dataset-name" class="dataset-label">Dataset name</label>
-        <VInputText v-model="datasetName" class="dataset-name" :autofocus="true" :inputId="'dataset-name'"
-                placeholder="My board game" />
-        <div class="dialog-controls">
-            <VButton outlined label="Cancel" @click="datasetDialogVisible = false" />
-            <VButton label="Submit" @click="handleSubmitDatasetClick" />
-        </div>
-    </VDialog>
-    <VDialog v-model:visible="compareDialogVisible" modal header="Select dataset" class="compare-dialog"
-            :dismissable-mask="true" :draggable="false">
-        <div class="compare-dataset-list">
-            <div v-for="(dataset, index) in userDatasets.sort((a, b) => b.timestamp - a.timestamp)" :key="index">
-                <DatasetListItemComponent v-bind="dataset" />
-                <VButton text label="Select" class="compare-select" @click="compareWithDataset(dataset.id)" />
-            </div>
-        </div>
-        <div class="dialog-controls">
-            <VButton outlined label="Cancel" @click="compareDialogVisible = false" />
-        </div>
-    </VDialog>
-</template>
-
-
-<style scoped>
-.quantities {
-    overflow: hidden;
-}
-
-.quantities-header {
-    display: flex;
-    font-size: 0.75rem;
-    font-weight: 300;
-    letter-spacing: 0.3px;
-    margin: 12px 0 6px 0;
-    color: var(--primary-color);
-    user-select: none;
-}
-
-.quantities-content {
-    max-height: 60vh;
-}
-
-.quantities-col:nth-child(1) {
-    flex-basis: 15%;
-    text-align: center;
-}
-
-.quantities-col:nth-child(2) {
-    flex-basis: 60%;
-    text-indent: 10px;
-}
-
-.quantities-col:nth-child(3) {
-    flex-basis: 25%;
-    text-align: center;
-}
-
-.quantities-label-notice {
-    font-size: 0.8rem;
-    line-height: 1.1rem;
-    font-weight: 300;
-    letter-spacing: 0.3px;
-    color: var(--text-color-secondary);
-    opacity: 0.7;
-    user-select: none;
-    text-align: center;
-}
-
-.image-view-tool-bar {
-    padding: 0;
-    position: fixed;
-    bottom: 0;
-    height: 90px;
-    align-items: stretch;
-}
-
-.image-view-tool-bar > button {
-    flex-direction: column;
-    padding: 12px 1rem;
-    justify-content: space-between;
-    flex: 1 1 0px;
-}
-
-.element-count {
-    display: flex;
-    flex-direction: column;
-    color: var(--primary-color);
-    text-align: center;
-    justify-content: space-between;
-    margin-top: -20px;
-    background-color: var(--surface-card);
-    border-top-left-radius: 30px;
-    border-top-right-radius: 30px;
-    padding: 10px 25px;
-    -webkit-box-shadow: 0px 0px 6px 0px var(--color-shadow);
-    -moz-box-shadow: 0px 0px 6px 0px var(--color-shadow);
-    box-shadow: 0px 0px 6px 0px var(--color-shadow);
-}
-
-.element-count-value {
-    display: block;
-    font-weight: 700;
-    font-size: 2.25rem;
-}
-
-.element-count-label {
-    display: block;
-    font-weight: 500;
-}
-
-.rename-notice,
-.no-elements-notice {
-    margin-left: 10px;
-    color: var(--text-color-secondary);
-    opacity: 0.7;
-}
-
-.no-elements-notice {
-    margin: 20px 0 10px 0;
-    text-align: center;
-}
-
-.dataset-dialog input {
-    margin-bottom: 30px;
-    width: 100%;
-}
-
-.dataset-dialog p {
-    user-select: none;
-    line-height: 1.8rem;
-}
-
-.dataset-dialog span {
-    font-weight: 600;
-    font-size: 1.4rem;
-    margin-right: 2px;
-}
-
-.dataset-dialog label {
-    display: inline-block;
-    margin-bottom: 8px;
-    margin-top: 16px;
-    color: var(--text-color-secondary);
-    user-select: none;
-}
-
-.dataset-dialog .dialog-controls,
-.compare-dialog .dialog-controls {
-    display: flex;
-    justify-content: flex-end;
-    gap: 12px;
-}
-
-.compare-button {
-    position: fixed;
-    bottom: 130px;
-    left: 50%;
-    transform: translateX(-50%);
-    max-width: 240px;
-    width: 100%;
-}
-
-.compare-dialog .compare-dataset-list {
-    max-height: 65vh;
-    overflow-y: auto;
-    margin-bottom: 18px;
-}
-
-.compare-dialog .compare-dataset-list > div {
-    position: relative;
-}
-
-.compare-dialog .compare-dataset-list > div:not(:last-child) {
-    border-bottom: 1px solid var(--surface-border);
-}
-
-.inactive-button {
-    pointer-events: none;
-    opacity: 0.5;
-}
-
-.compare-select {
-    position: absolute;
-    top: 50%;
-    right: 6px;
-    transform: translateY(-50%);
-}
-</style>
-
-<style>
-.compare-dialog {
-    width: 92.5vw;
-}
-
-.quantities .p-sidebar-header {
-    padding-bottom: 12px;
-}
-
-.quantities .p-sidebar-header-content {
-    color: var(--primary-color);
-    font-weight: 400;
-    letter-spacing: 0.3px;
-    user-select: none;
-}
-
-.quantity-switch .p-inputswitch-input {
-    width: 40px;
-    left: 50%;
-    transform: translateX(-50%);
-}
-
-.quantity-switch .p-inputswitch-slider {
-    width: 40px;
-    margin: 0 auto;
-}
-</style>
